backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sendgrid==6.10.0
azure-storage-blob==12.19.0
openai-whisper==20231117
transformers==4.36.0
torch==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
sqlalchemy==2.0.23
redis==5.0.1